fix(jobs): fix job instancing issue
Sometimes, a group may not have a character set or data related to the character may be removed
from the database due to expired token or other things. With the current way, when jobs are
instancing, it may throw an exception. Since no try/catch has been put on the dispatcher, it will
crash and stop queuing further jobs. This commit is avoiding exception on the GroupSync job
constructor and if no main has been set, attempt to continue the process using the first character
attached to the group. Also, in order to avoid strange pattern, funnel is based on group ID instead
main_character.